restructure: make the directory structure more intuitive.
