+++ bots/event_trading_framework/helpers/transaction_manager.py
