+++ bots/example_bots/__init__.py
@@ -1,0 +1,2 @@
+from .types import GridParams
+from .gridbot import GridBot
