+++ bots/example_bots/types/__init__.py
@@ -1,0 +1,1 @@
+from .grid_params import GridParams
