+++ bots/example_bots/types/types.py
