+++ bots/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "bots"
-version = "0.1.0"
+version = "0.1.1"
 description = "A python trading framwork using the rubi-py client library."
 authors = ["adam <adam@rubicon.finance>"]
 readme = "README.md"
@@ -8,11 +8,11 @@ readme = "README.md"
 [tool.poetry.dependencies]
 python = "^3.10"
 python-dotenv = "1.0.0"
-rubi = "2.0.3"
+#rubi = "2.0.4"
 
 
 [tool.poetry.group.dev.dependencies]
-#rubi = { path = "../rubi", develop = true }
+rubi = { path = "../rubi", develop = true }
 
 [build-system]
 requires = ["poetry-core"]
